REFACTOR: 필요없는 import 문 제거 및 Empty View 디자인 수정
